student_proj/requirements.txt
Django<3
Pillow==9.1.0
psycopg2-binary==2.9
django-crispy-forms==1.9.1